requirements.txt
# Telegram bot dependencies
python-telegram-bot>=20.7
aiogram>=3.0.0
aiolimiter>=1.1.0
aiocache>=0.12.0
python-dotenv>=1.0.0
loguru>=0.7.0
pydantic[email]>=2.0.0
Pillow>=10.0.0
validators>=0.22.0
aiohttp>=3.8.0
requests>=2.31.0
redis>=4.5.0

# Dev dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
black>=23.0.0
ruff>=0.1.0